% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corclust.R
\name{corclust}
\alias{corclust}
\title{Clustering Correlations from Hierarchical Clustering Tree}
\usage{
corclust(x, y = NULL, k = NULL)
}
\arguments{
\item{x}{a matrix}

\item{y}{a vector}

\item{k}{an integer}
}
\description{
Clustering Correlations from Hierarchical Clustering Tree
}
\author{
David Senhora Navega
}
