% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.R
\name{cramer.test}
\alias{cramer.test}
\title{Compute Cramer' Phi (V) between two factors}
\usage{
cramer.test(x, y, bias = T, ...)
}
\arguments{
\item{x}{a factor}

\item{y}{a factor}

\item{bias}{a logical stating if bias correction applied. Default is TRUE.}

\item{...}{additional arguments to \code{\link[stats]{chisq.test}}}
}
\value{
an htest object
}
\description{
Compute Cramer' Phi (V) between two factors
}
\author{
David Senhora Navega
}
