% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{eta}
\alias{eta}
\title{Compute Pearson's Eta between a data.frame of factors and a numeric vector}
\usage{
eta(x, y)
}
\arguments{
\item{x}{a data.frame where all columns are factors.}

\item{y}{a numeric vector.}
}
\value{
a vector of eta values
}
\description{
Compute Pearson's Eta between a data.frame of factors and a numeric vector
}
\author{
David Senhora Navega
}
