% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.R
\name{etasq.test}
\alias{etasq.test}
\title{Compute Pearson's Eta Squared between a factor (x) and a numeric (y) vector.}
\usage{
etasq.test(x, y)
}
\arguments{
\item{x}{a factor vector}

\item{y}{a numeric vector}
}
\value{
an htest object
}
\description{
Compute Pearson's Eta Squared between a factor (x) and a numeric (y) vector.
}
\author{
David Senhora Navega
}
