% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbf.R
\name{fcbf}
\alias{fcbf}
\title{Fast Correlation-Based Filter for Feature Selection}
\usage{
fcbf(x, y)
}
\arguments{
\item{x}{a symmetric numeric matrix with inter-correlation (association) of
the features (input) under study. See Details.}

\item{y}{a vector with the correlation (association) of x (features) with the
class (output). See Details}
}
\value{
a logical vector of selected features
}
\description{
Fast Correlation-Based Filter for Feature Selection
}
\details{
In this implementation, and following Hall (1999), correlation is used its
more general sense. Is up to the user to compute the matrix x and vector y
using the correlation / association measure that best suite the data and
problem at hand.
}
\references{
Hall, M.A., 1999. Correlation-based feature selection for machine learning.

Yu, L. and Liu, H., 2003. Feature selection for high-dimensional data:
A fast correlation-based filter solution. In Proceedings of the
20th international conference on machine learning (ICML-03) (pp. 856-863).
}
\author{
David Senhora Navega
}
