% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcbw.R
\name{fcbw}
\alias{fcbw}
\title{Fast Computation of Pearson and Spearman Correlation Coefficient}
\usage{
fcbw(
  x,
  y,
  method = c("pearson", "spearman"),
  partial = T,
  weights = NULL,
  digits = 3
)
}
\arguments{
\item{x}{a matrix or data.frame. Data is internally coerced to numeric
representation.}

\item{y}{vector (numeric or factor). Internally coerced to numeric
representation.}

\item{method}{a character stating the method to used. Default is "pearson".}

\item{partial}{a logical stating if correlation among x should be account
for y. Default is T.}

\item{weights}{a numeric vector of weights used to computed weighted
correlations.}

\item{digits}{an integer defining the precision of round function. Default is
3}
}
\value{
a list with two components, x and y, storing the inter-correlations
among x and the correlations of x with y respectively. The component x
is a symmetric squared matrix and the component y is a numeric vector.
}
\description{
Fast Computation of Pearson and Spearman Correlation Coefficient
}
\details{
This implementation uses matrix operations for fast computation of correlation
coefficients. A drawback of this approach is that the data must be complete,
that is no NA values. This function implements a fast mean imputation step
where the NA values are substituted by 0 after the data been scaled as part of
the algorithm for fast computation of the correlation.
}
\author{
David Senhora Navega
}
