% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gafs.R
\name{gafs.control}
\alias{gafs.control}
\title{Genetic Algorithm Feature Selection Parameters}
\usage{
gafs.control(
  npopulation = 100,
  ngeneration = 200,
  pressure = 0.8,
  crossover = 0.8,
  mutation = 0.05,
  elitism = 0.05,
  warmup = 0.5,
  early = 0.05,
  adaptive = T,
  rate = 0.01,
  alpha = 0.01,
  digits = 3
)
}
\arguments{
\item{npopulation}{number of individuals in the initial population.}

\item{ngeneration}{number of generation}

\item{pressure}{selective pressure probability. Probability of selecting the
best individual in a binary tournament selection}

\item{crossover}{crossover probability.}

\item{mutation}{mutation probability.}

\item{elitism}{fraction of best individuals that are transferred directly to
the next generation.}

\item{warmup}{fraction of generations at which early stop and adaptive
behavior will trigger.}

\item{early}{fraction of consecutive generations at which the GA will be
terminated if no improvement is detected to the fitness function.}

\item{adaptive}{a logical stating if probabilities defined by pressure,
crossover and mutation should be adapted according to an exponential function
after the warm-up phase. Default is TRUE.}

\item{rate}{the rate of the exponential function to adapt the parameters of
the GA. Default is 0.01 (i.e. probability will increase or decrease by 1\%)}

\item{alpha}{level of significance of binomial test performed to assess
if a feature should be accepted or rejected based the the best individual
chromosomes of all generations. Default is 0.01}

\item{digits}{an integer defining the precision of the round function.
Default is 3.}
}
\description{
Genetic Algorithm Feature Selection Parameters
}
\author{
David Senhora Navega
}
