% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gafs.R
\name{gafs}
\alias{gafs}
\title{Genetic Algorithm Feature Selection Parameters}
\usage{
gafs(x, y, control = gafs.control())
}
\arguments{
\item{x}{a symmetric numeric matrix with inter-correlation (association) of
the features (input) under study. See Details.}

\item{y}{a vector with the correlation (association) of x (features) with the
class (output). See Details}

\item{control}{a list of parameters for genetic algorithm as created by
\code{\link{gafs.control}}}
}
\value{
a gafs object
}
\description{
Genetic Algorithm Feature Selection Parameters
}
\details{
In this implementation, and following Hall (1999), correlation is used its
more general sense. Is up to the user to compute the matrix x and vector y
using the correlation / association measure that best suite the data and
problem at hand.
}
\references{
Hall, M.A., 1999. Correlation-based feature selection for machine learning.
}
\author{
David Senhora Navega
}
