% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcortree.R
\name{ggcortree}
\alias{ggcortree}
\title{Visualize a Correlation Matrix with an Hierarchical Tree using ggplot2}
\usage{
ggcortree(
  x,
  y = NULL,
  k = NULL,
  circular = F,
  nudge = 0.05,
  expand = 0.2,
  label.size = 3.5
)
}
\arguments{
\item{x}{a matrix of pairwise correlation among features}

\item{y}{vector of correlation of the features represented in x with a target
quantity. Default is NULL which means that this information is not exploited.
If y is supplied it is used to reorder the hierarchical tree under the
constrains of its structure.}

\item{k}{an integer. Number of cluster to cut the tree into. Default is NULL.
k is estimate as the minimal number of clusters the would result in an isolated
feature.}

\item{circular}{should the tree be represent in polar coordinates? Default F}

\item{nudge}{a numeric.  Displacement factor of the labels. Aesthetics purpose.}

\item{expand}{a numeric. Expansion of the y-axis. Aesthetics purpose.}

\item{label.size}{a numeric. Size of the terminal nodes of the tree labels.}
}
\description{
Visualize a Correlation Matrix with an Hierarchical Tree using ggplot2
}
\author{
David Senhora Navega
}
