% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.R
\name{phi}
\alias{phi}
\title{Compute association matrix between factor variables using Cramer's Phi (V)}
\usage{
phi(x)
}
\arguments{
\item{x}{a data.frame where all columns are factors.}
}
\value{
a symmetric matrix with the pairwise phi values among the columns of
x
}
\description{
Compute association matrix between factor variables using Cramer's Phi (V)
}
\author{
David Senhora Navega
}
